Fixed mistake where it accesed wrong table
