+++ style.css
@@ -5,3 +5,13 @@
 
 .block { border:  1px solid black; background-color: gray; width: 50px; font-size: 12px; text-align: center;}
 .start { width: 480px; height: 50px; margin: 0px; border: 1px solid black; background-color: gray; position: absolute;}
+.box {
+    float: left;
+    width: 225px;
+    height:300px;
+    margin-left: 10px;
+    margin-top: 10px;
+    padding: 5px;
+    overflow:scroll;
+    border: 2px black;
+}
